Add output of git diff to the app
The output of git diff is stored as one giant string in the state, and
then is splashed on the diff panel when the app is rendered. In the
future, this will only display the diff for the currently highlighted
file in the staged/unstaged changes panel (and perhaps a diff of the
highlighted commit in the commit log panel).